Refactored: Replace while loop with java stream

+++ src/test/java/censusanalyser/CensusAnalyserTest.java
@@ -13,8 +13,8 @@ public class CensusAnalyserTest {
     public void givenIndianCensusCSVFileReturnsCorrectRecords() {
         try {
             CensusAnalyser censusAnalyser = new CensusAnalyser();
-            int numOfRecords = censusAnalyser.loadIndiaCensusData( INDIA_CENSUS_CSV_FILE_PATH );
-            Assert.assertEquals( 29, numOfRecords );
+            Integer numOfRecords = censusAnalyser.loadIndiaCensusData( INDIA_CENSUS_CSV_FILE_PATH );
+            Assert.assertEquals( ( Integer ) 29, numOfRecords );
         } catch ( CensusAnalyserException e ) {
             e.printStackTrace();
         }
